% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage Data}
\format{
A data frame with 100 rows and the following columns:
\describe{
  \item{Distance}{Distance in miles}
  \item{Damage}{Damage in million of dollars}
}
}
\source{
MATH 4753 resources
}
\usage{
fire
}
\description{
A dataset containing information on fire damage for analysis purposes.
}
\keyword{datasets}
