% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Compute and Plot Normal Curve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Mean of the normal distribution}

\item{sigma}{Standard deviation of the normal distribution}

\item{a}{Upper limit for shading and probability calculation}
}
\value{
A list with \code{mu}, \code{sigma}, and \code{probability}
}
\description{
Compute and Plot Normal Curve
}
\examples{
mycurve <- myncurve(10, 5, 6)
print(mycurve)

}
