% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Calculates the number of  tickets to sell to minimize overbooking}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Integer, number of seats availible}

\item{gamma}{Numeric, acceptable probability threshold for overbooking}

\item{p}{Numberic, the probability that a ticket holder will attend}
}
\value{
A list containing:
\describe{
   \item{nd}{The number of tickets to sell using the discrete distribution.}
   \item{nc}{The number of tickets to sell using the normal distribution.}
   \item{N}{The number of seats avalible.}
   \item{p}{The proability that a ticket holder will attend their flight}
   \item{gamma}{the acceptable probability threshold for overbooking}
}
}
\description{
Calculates the number of  tickets to sell to minimize overbooking
}
\examples{
\dontrun{
ntickets(N = 100, gamma = 0.5, p = 0.9)
}
}
